+++ src_smartcontract_db/transaction_log_alter_modify/AlterDropPrimaryKeyCommandLog.cpp
@@ -24,7 +24,7 @@ void AlterDropPrimaryKeyCommandLog::setCommand(AlterDropPrimaryKeyCommand* comma
 }
 
 int AlterDropPrimaryKeyCommandLog::binarySize() const {
-	int total = sizeof(uint16_t);
+	int total = sizeof(uint8_t);
 
 	return total;
 }
