+++ src_smartcontract_db/transaction_log_alter_modify/AlterRenameTableCommandLog.cpp
@@ -24,7 +24,7 @@ void AlterRenameTableCommandLog::setCommand(AlterRenameTableCommand* command) no
 }
 
 int AlterRenameTableCommandLog::binarySize() const {
-	int total = sizeof(uint16_t);
+	int total = sizeof(uint8_t);
 
 	return total;
 }
